concepts/tut_aws/c1-rep/cdk-app-python-project/requirements.txt
aws-cdk-lib==2.0.0
constructs>=10.0.0
Jinja2==3.0.0
requests==2.25.1
